+++ nkululeko/feat_extract/feats_import.py
@@ -30,7 +30,7 @@ class ImportSet(Featureset):
                 "feature type == import needs import_file = ['file1', 'filex']"
             )
         except SyntaxError:
-            if type(feat_import_files) == str:
+            if type(feat_import_files) is str:
                 feat_import_files = [feat_import_files]
             else:
                 self.util.error(f"import_file is wrong: {feat_import_files}")
@@ -40,6 +40,11 @@ class ImportSet(Featureset):
             if not os.path.isfile(feat_import_file):
                 self.util.error(f"no import file: {feat_import_file}")
             df = audformat.utils.read_csv(feat_import_file)
+            if df.isnull().values.any():
+                self.util.warn(
+                    f"imported features contain {df.isna().sum()} NAN, filling with zero."
+                )
+                df = df.fillna(0)
             df = self.util.make_segmented_index(df)
             df = df[df.index.isin(self.data_df.index)]
             if import_files_append: